+++ varfish/static/js/filter_form.js
@@ -923,7 +923,7 @@ const presets = {
       "quality-field-dp-het": 10,
       "quality-field-dp-hom": 5,
       "quality-field-ab": 0.2,
-      "quality-field-gq": 20,
+      "quality-field-gq": 10,
       "quality-field-ad": 3,
       "quality-field-fail": "drop-variant",
       "quality-field-ad-max": null,